Take test_security out of test_ecr_scan and increase the enhanced scanning timeout
