+++ test/dlc_tests/sanity/test_ecr_scan.py
@@ -44,27 +44,6 @@ from src.config import is_ecr_scan_allowlist_feature_enabled
 ALLOWLIST_FEATURE_ENABLED_IMAGES = {"mxnet": SpecifierSet(">=1.8.0,<1.9.0")}
 
 
-@pytest.mark.usefixtures("sagemaker")
-@pytest.mark.model("N/A")
-@pytest.mark.canary("Run security test regularly on production images")
-def test_security(image):
-    repo_name, image_tag = image.split("/")[-1].split(":")
-    container_name = f"{repo_name}-{image_tag}-security"
-
-    run(
-        f"docker run -itd --name {container_name} "
-        f"--mount type=bind,src=$(pwd)/container_tests,target=/test"
-        f" --entrypoint='/bin/bash' "
-        f"{image}",
-        echo=True,
-    )
-    try:
-        docker_exec_cmd = f"docker exec -i {container_name}"
-        run(f"{docker_exec_cmd} python /test/bin/security_checks.py ")
-    finally:
-        run(f"docker rm -f {container_name}", hide=True)
-
-
 def is_image_covered_by_allowlist_feature(image):
     """
     This method checks if the allowlist feature has been enabled for the image
