+++ test/test_utils/security.py
@@ -813,7 +813,7 @@ def wait_for_enhanced_scans_to_complete(ecr_client, image):
     scan_status = None
     scan_status_description = ""
     start_time = time()
-    while (time() - start_time) <= 600:
+    while (time() - start_time) <= 1200:
         try:
             scan_status, scan_status_description = ecr_utils.get_ecr_image_enhanced_scan_status(ecr_client, image)
         except ecr_client.exceptions.ScanNotFoundException as e:
